refactor: `WritingViewer`에서 `WritingTitle` 분리
